Allow using the latest version of PHPUnit
While it does not mean that we can use the latest features, it means we
can spot and address deprecations earlier. It also means the output of
composer outdated is smaller, which in turns allows us to spot where
work is needed more easily.
Note that it also means running phpunit now also produces a warning
about validating against a deprecated schema, and I do not think we can
address it until we drop support for PHP <= 8.1, since PHPUnit 10
requires PHP 8.1+